+++ src/cli/commands/public-cmds/status-cmd.js
@@ -28,11 +28,13 @@ export default class Status extends Command {
   name = 'status';
   description = `show the working area component(s) status.\n  https://${BASE_DOCS_DOMAIN}/docs/cli-status.html`;
   alias = 's';
-  opts = [];
+  opts = [['j', 'json', 'return a json version of the component']];
   loader = true;
   migration = true;
+  json = false;
 
-  action(): Promise<Object> {
+  action([], { json }: { json?: boolean }): Promise<Object> {
+    this.json = json;
     return status();
   }
 
@@ -46,6 +48,22 @@ export default class Status extends Command {
     invalidComponents,
     outdatedComponents
   }: StatusResult): string {
+    if (this.json) {
+      return JSON.stringify(
+        {
+          newComponents,
+          modifiedComponent: modifiedComponent.map(c => c.id.toString()),
+          stagedComponents: stagedComponents.map(c => c.id()),
+          componentsWithMissingDeps: componentsWithMissingDeps.map(c => c.id.toString()),
+          importPendingComponents: importPendingComponents.map(c => c.id.toString()),
+          autoTagPendingComponents,
+          invalidComponents,
+          outdatedComponents: outdatedComponents.map(c => c.id.toString())
+        },
+        null,
+        2
+      );
+    }
     // If there is problem with at least one component we want to show a link to the
     // troubleshooting doc
     let showTroubleshootingLink = false;