+++ src/consumer/consumer.js
@@ -57,6 +57,7 @@ import type { InvalidComponent } from './component/consumer-component';
 import MainFileRemoved from './component/exceptions/main-file-removed';
 import type { BitIdStr } from '../bit-id/bit-id';
 import ExtensionFileNotFound from '../extensions/exceptions/extension-file-not-found';
+import { getAutoTagPending } from '../scope/component-ops/auto-tag';
 
 type ConsumerProps = {
   projectPath: string,
@@ -335,12 +336,7 @@ export default class Consumer {
           'load components',
           `failed loading ${Analytics.hashData(id.toString())} from the file-system`
         );
-        if (
-          err instanceof MissingFilesFromComponent ||
-          err instanceof ComponentNotFoundInPath ||
-          err instanceof ExtensionFileNotFound ||
-          err instanceof MainFileRemoved
-        ) {
+        if (Component.isComponentInvalidByErrorType(err)) {
           invalidComponents.push({ id, error: err });
           return null;
         }
@@ -395,22 +391,22 @@ export default class Consumer {
     return !this.bitJson.distEntry && !this.bitJson.distTarget;
   }
 
-  async candidateComponentsForAutoTagging(modifiedComponents: BitId[]): Promise<?(BitId[])> {
+  async potentialComponentsForAutoTagging(modifiedComponents: BitIds): Promise<BitIds> {
     const candidateComponents = this.bitMap.getAllBitIds([COMPONENT_ORIGINS.AUTHORED, COMPONENT_ORIGINS.IMPORTED]);
-    if (!candidateComponents) return null;
     const modifiedComponentsWithoutVersions = modifiedComponents.map(modifiedComponent =>
       modifiedComponent.toStringWithoutVersion()
     );
     // if a modified component is in candidates array, remove it from the array as it will be already tagged with the
     // correct version
-    return candidateComponents.filter(
+    const idsWithoutModified = candidateComponents.filter(
       component => !modifiedComponentsWithoutVersions.includes(component.toStringWithoutVersion())
     );
+    return BitIds.fromArray(idsWithoutModified);
   }
 
-  async listComponentsForAutoTagging(modifiedComponents: BitId[]): Promise<ModelComponent[]> {
-    const candidateComponents = await this.candidateComponentsForAutoTagging(modifiedComponents);
-    return this.scope.bumpDependenciesVersions(candidateComponents, modifiedComponents, false);
+  async listComponentsForAutoTagging(modifiedComponents: BitIds): Promise<ModelComponent[]> {
+    const candidateComponents = await this.potentialComponentsForAutoTagging(modifiedComponents);
+    return getAutoTagPending(this.scope, candidateComponents, modifiedComponents);
   }
 
   /**